# docs/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "StatsTid Backend API",
    "version": "v1"
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "StatsTid.Backend.Api"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatsTid.Backend.Api.Contracts.LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsTid.Backend.Api.Contracts.LoginResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "StatsTid.Backend.Api.Contracts.LoginRequest": {
        "required": [
          "password",
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "StatsTid.Backend.Api.Contracts.LoginResponse": {
        "required": [
          "employeeId",
          "expiresAt",
          "orgId",
          "role",
          "token"
        ],
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "employeeId": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "orgId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}